chore: fix semantic release by using custom module
